remove replica_index from replicas and replica_under_construction tables
read invalidated blocks by storageId
