+++ src/main/java/io/hops/metadata/hdfs/entity/CorruptReplica.java
@@ -17,7 +17,7 @@ package io.hops.metadata.hdfs.entity;
 
 import io.hops.metadata.common.FinderType;
 
-public class CorruptReplica extends Replica {
+public class CorruptReplica extends ReplicaBase {
 
   public static enum Finder implements FinderType<CorruptReplica> {
